Code cleanup

+++ scs/src/ConcretePriceVisitor.java
@@ -1,10 +1,10 @@
 public class ConcretePriceVisitor implements PriceVisitor {
 	
-	public float visit(UnitPricedLineItem uli) {
+	public double visit(UnitPricedLineItem uli) {
 		return uli.getPricePerUnit() * uli.getQuantity();
 	}
 	
-	public float visit(WeightPricedLineItem wli) {
+	public double visit(WeightPricedLineItem wli) {
 		return wli.getPricePerKg() * wli.getKgWeight();
 	}
 }
