+++ scs/src/ConcreteWeightVisitor.java
@@ -1,10 +1,10 @@
 public class ConcreteWeightVisitor implements WeightVisitor {
 	
-	public float visit(UnitPricedLineItem uli) {
+	public double visit(UnitPricedLineItem uli) {
 		return uli.getKgPerUnit() * uli.getQuantity();
 	}
 	
-	public float visit(WeightPricedLineItem wli) {
+	public double visit(WeightPricedLineItem wli) {
 		return wli.getKgWeight();
 	}
 }
