+++ scs/src/PriceVisitor.java
@@ -1,6 +1,6 @@
 public interface PriceVisitor {
 	// Visitor pattern
 	// Calcule le prix d'une instance d'une sous-classe concrète de LineItem
-	public float visit(UnitPricedLineItem uli);
-	public float visit(WeightPricedLineItem wli);
+	public double visit(UnitPricedLineItem uli);
+	public double visit(WeightPricedLineItem wli);
 }
