+++ scs/src/PricingPolicy.java
@@ -1,5 +1,5 @@
 public interface PricingPolicy {
 	// Strategy pattern
 	// Permet de librement définir de nouvelles politiques tarifaires.
-	public float applyDiscount(float price);
+	public double applyDiscount(double price);
 }
