+++ scs/src/WeightPricedLineItem.java
@@ -1,26 +1,26 @@
 public class WeightPricedLineItem extends LineItem {
 	private WeightPricedItem item_;
-	private float kgWeight_;
+	private double kgWeight_;
 	
-	public WeightPricedLineItem(WeightPricedItem item, float kgWeight, PricingPolicy policy) {
+	public WeightPricedLineItem(WeightPricedItem item, double kgWeight, PricingPolicy policy) {
 		super(policy);
 		this.item_ = item;
 		this.kgWeight_ = kgWeight;
 	}
 	
-	public float getPricePerKg() {
+	public double getPricePerKg() {
 		return this.item_.getPricePerKg();
 	}
 	
-	public float getKgWeight() {
+	public double getKgWeight() {
 		return this.kgWeight_;
 	}
 	
-	public float accept(PriceVisitor pv) {
+	public double accept(PriceVisitor pv) {
 		return pv.visit(this);
 	}
 	
-	public float accept(WeightVisitor wv) {
+	public double accept(WeightVisitor wv) {
 		return wv.visit(this);
 	}
 }
